+++ src/Category_Colors/Frontend.php
@@ -72,8 +72,8 @@ class Frontend {
 
 		// Optionally add legend superpowers
 		if ( isset( $this->options['legend_superpowers'] ) &&
-			 '1' === $this->options['legend_superpowers'] &&
-			 ! wp_is_mobile()
+			'1' === $this->options['legend_superpowers'] &&
+			! wp_is_mobile()
 		) {
 			wp_enqueue_script( 'legend_superpowers', TECCC_RESOURCES . '/legend-superpowers.js', array( 'jquery' ), Main::$version, true );
 		}