+++ CHANGES.md
@@ -46,7 +46,7 @@
 * graceful failure if The Events Calendar is not active
 
 #### 4.6.2 / 2017-01-20
-* added CSS selector to override Customizer month view 
+* added CSS selector to override Customizer month view
 
 #### 4.6.1 / 2017-01-19
 * added filters and functions to provide WPML compatibility thanks @jvier for testing