+++ CHANGES.md
@@ -1,6 +1,7 @@
 #### [unreleased]
 * update for new tooltip title as `h3`
 * added some functions for adding deprecated CSS
+* added function `Extras::override_customizer()` for when _Customizer_ settings need to be overridden
 
 #### 5.0.1 / 2018-04-06
 * needed a space between CSS parameters, PhpStorm reformat code error